fix: pipeline all-paid detection — cast COUNT/SUM to int/numeric in SQL, safe Number() comparison in component

+++ src/app/api/events/route.ts
@@ -28,9 +28,9 @@ export async function GET(request: NextRequest) {
     const offset = parseInt(searchParams.get('offset') ?? '0', 10) || 0;
 
     let query = `SELECT e.*,
-      (SELECT COALESCE(SUM(amount), 0) FROM payments WHERE event_id = e.id AND paid = true) as total_paid,
-      (SELECT COUNT(*) FROM payments WHERE event_id = e.id AND paid = false) as pending_payments,
-      (SELECT COUNT(*) FROM payments WHERE event_id = e.id) as total_payments
+      (SELECT COALESCE(SUM(amount), 0)::numeric FROM payments WHERE event_id = e.id AND paid = true) as total_paid,
+      (SELECT COUNT(*)::int FROM payments WHERE event_id = e.id AND paid = false) as pending_payments,
+      (SELECT COUNT(*)::int FROM payments WHERE event_id = e.id) as total_payments
     FROM events e`
     const params: any[] = [];
     const conditions: string[] = [];
